+++ src/main/java/com/selfxdsd/storage/SelfTasks.java
@@ -424,9 +424,7 @@ public final class SelfTasks implements Tasks {
                 rec.getValue(SLF_PMS_XDSD.USERNAME),
                 rec.getValue(SLF_PMS_XDSD.PROVIDER),
                 rec.getValue(SLF_PMS_XDSD.ACCESS_TOKEN),
-                BigDecimal.valueOf(
-                    rec.getValue(SLF_PMS_XDSD.COMMISSION).longValue()
-                ),
+                rec.getValue(SLF_PMS_XDSD.COMMISSION).doubleValue(),
                 this.storage
             ),
             this.storage
